requirements.txt
fastapi
uvicorn
sentence-transformers
numpy

python-dotenv
